requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
requests==2.31.0
beautifulsoup4==4.12.3
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
scikit-learn==1.4.0
APScheduler==3.10.4
python-dotenv==1.0.0
gunicorn==21.2.0
